split pluginManager/engineManager/windowManager from ImJoy.vue
